+++ NEWS.md
@@ -6,6 +6,8 @@
   data frame.
 * `as_Workbook()` is now much faster when adding many tables to an Excel
   workbook. Thanks @lemonad for the report.
+* New table property `breakable()` allows tables to break between rows across
+  pages in HTML, LaTeX, RTF, Typst and Word output.
   
 # huxtable 5.8.0
 
@@ -789,5 +791,4 @@ For example, the following code now does what you probably want:
 
 * Added a `NEWS.md` file to track changes to the package.
 * First CRAN release.
-